Change a lot of data types into float (from double) in DataSource

+++ Source/Gauges/GenericBargraph.h
@@ -23,7 +23,7 @@
 namespace OpenGC
 {
 
-typedef double (AirframeDataContainer::*AirframeMemFn)();
+typedef float (AirframeDataContainer::*AirframeMemFn)();
 
 class GenericBargraph : public GaugeComponent
 {
