+++ Source/Gauges/PieDial.h
@@ -24,7 +24,7 @@
 namespace OpenGC
 {
 
-typedef double (AirframeDataContainer::*AirframeMemFn)();
+typedef float (AirframeDataContainer::*AirframeMemFn)();
 
 class PieDial : public MarkedDial
 {
